refactored blockchain updation listener package
